requirements.txt
fastapi
uvicorn
transformers
tensorflow
opencv-python
peft
cohere
pydantic
python-multipart
numpy
pillow
openai-whisper
tensorflow[and-cuda]
assemblyai
vosk
pyttsx3
gtts


